refactor: remove return repository, move IO from handlers, improve CLI mappers
- Removed SnapshotReturnRepository and switched return listing to common order repository via OrdersFilterRequest with status=RETURNED
- Return listing is now sorted by UpdatedStatusAt
- Removed all stdin/stdout logic from handlers; CLI layer now handles all user interaction
- Refactored CLI param mappers (list, scroll, etc.) for clarity and correctness
- Cleaned up filter application logic in repository (UserID now optional via pointer)
- Write first version of gateway error handler

+++ pvz-cli/internal/grpc/mappers/shared_mappings.go
@@ -15,8 +15,8 @@ func toPbOrder(o models.Order) *pb.Order {
 		UserId:     o.UserID,
 		Status:     toPbOrderStatus(o.Status),
 		ExpiresAt:  timestamppb.New(o.ExpiresAt),
-		Weight:     float32(o.Weight),
-		TotalPrice: float32(o.Price),
+		Weight:     o.Weight,
+		TotalPrice: o.Price,
 		Package:    toPbPackageTypePtr(o.Package),
 	}
 }
