+++ pvz-cli/internal/usecases/services/package_pricing_service.go
@@ -4,5 +4,5 @@ import "pvz-cli/internal/models"
 
 // PackagePricingService calculates package pricing and validates weight constraints
 type PackagePricingService interface {
-	Evaluate(pkg models.PackageType, weight, price float64) (surcharge float64, err error)
+	Evaluate(pkg models.PackageType, weight, price float32) (surcharge float32, err error)
 }
