+++ pvz-cli/internal/usecases/services/strategies/pricing_strategy.go
@@ -4,5 +4,5 @@ import "pvz-cli/internal/models"
 
 // PricingStrategy defines the interface for calculating a surcharge based on the selected package type.
 type PricingStrategy interface {
-	GetSurcharge(pkg models.PackageType) float64
+	GetSurcharge(pkg models.PackageType) float32
 }
