+++ pvz-cli/internal/usecases/services/validators/package_validator.go
@@ -4,5 +4,5 @@ import "pvz-cli/internal/models"
 
 // PackageValidator validates package types and weight constraints
 type PackageValidator interface {
-	Validate(pkg models.PackageType, weight float64) error
+	Validate(pkg models.PackageType, weight float32) error
 }
